pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nsscheduler"
version = "0.0.1"
authors = [
  { name=""},
]
description = ""
readme = "README.md"
requires-python = ">=3.11.1"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11"
]
dependencies = [
    "dash",
    "dash-bootstrap-components",
    "fastapi",
    "gevent",
    "pydantic",
    "pytz",
    "python-dateutil",
    "pyyaml",
    "uvicorn",
    "requests",
    "kubernetes",
    "awscli"
]
[project.scripts]
ns-scheduler = "nsscheduler.main:main"
ns-dashboard = "nsscheduler.dashboard.app:main"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-timeout"
]

[tool.setuptools]
packages = ["nsscheduler"]

[tool.black]
line-length = 120
